Move mandarin quiz to /mandarin/quiz
