+++ app/(language)/mandarin/quiz/cards/mandarin-match-pinyin.tsx
@@ -1,5 +1,5 @@
-import { MatchCard } from "../../quiz/match-cards";
-import { CardProps } from "../../quiz/types/card";
+import { MatchCard } from "../../../quiz/match-cards";
+import { CardProps } from "../../../quiz/types/card";
 import { MandarinMatchPinyinCardParams } from "../types/card";
 import { MandarinDefinition } from "../types/mandarin";
