+++ app/(language)/mandarin/quiz/utils/definition-utils.ts
