+++ app/(language)/mandarin/quiz/utils/generate-card-type.ts
